+++ com.unity.netcode.gameobjects/Runtime/Transports/UTP/UnityTransport.cs
@@ -17,6 +17,7 @@ using Unity.Burst;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Collections;
 using Unity.Jobs;
+using Unity.Netcode.Runtime;
 using Unity.Networking.Transport;
 using Unity.Networking.Transport.Relay;
 using Unity.Networking.Transport.Utilities;
@@ -43,7 +44,7 @@ namespace Unity.Netcode.Transports.UTP
         /// <param name="unreliableFragmentedPipeline">The UnreliableFragmented NetworkPipeline</param>
         /// <param name="unreliableSequencedFragmentedPipeline">The UnreliableSequencedFragmented NetworkPipeline</param>
         /// <param name="reliableSequencedPipeline">The ReliableSequenced NetworkPipeline</param>
-        void CreateDriver(
+        public void CreateDriver(
             UnityTransport transport,
             out NetworkDriver driver,
             out NetworkPipeline unreliableFragmentedPipeline,
@@ -106,6 +107,7 @@ namespace Unity.Netcode.Transports.UTP
     /// Note: This is highly recommended to use over UNet.
     /// </summary>
     [AddComponentMenu("Netcode/Unity Transport")]
+    [HelpURL(HelpUrls.UnityTransport)]
     public partial class UnityTransport : NetworkTransport, INetworkStreamDriverConstructor
     {
         /// <summary>
@@ -866,7 +868,11 @@ namespace Unity.Netcode.Transports.UTP
             var mtu = 0;
             if (NetworkManager)
             {
-                var ngoClientId = NetworkManager.ConnectionManager.TransportIdToClientId(sendTarget.ClientId);
+                var (ngoClientId, isConnectedClient) = NetworkManager.ConnectionManager.TransportIdToClientId(sendTarget.ClientId);
+                if (!isConnectedClient)
+                {
+                    return;
+                }
                 mtu = NetworkManager.GetPeerMTU(ngoClientId);
             }
 
@@ -1042,7 +1048,10 @@ namespace Unity.Netcode.Transports.UTP
                         continue;
                     }
                     var transportClientId = NetworkManager.ConnectionManager.ClientIdToTransportId(ngoConnectionId);
-                    ExtractNetworkMetricsForClient(transportClientId);
+                    if (transportClientId.Item2)
+                    {
+                        ExtractNetworkMetricsForClient(transportClientId.Item1);
+                    }
                 }
             }
             else
@@ -1276,7 +1285,7 @@ namespace Unity.Netcode.Transports.UTP
 
             if (NetworkManager != null)
             {
-                var transportId = NetworkManager.ConnectionManager.ClientIdToTransportId(clientId);
+                var (transportId, _) = NetworkManager.ConnectionManager.ClientIdToTransportId(clientId);
 
                 var rtt = ExtractRtt(ParseClientId(transportId));
                 if (rtt > 0)
@@ -1288,55 +1297,33 @@ namespace Unity.Netcode.Transports.UTP
             return (ulong)ExtractRtt(ParseClientId(clientId));
         }
 
+        /// <summary>
+        /// Provides the <see cref="NetworkEndpoint"/> for the NGO client identifier specified.
+        /// </summary>
+        /// <remarks>
+        /// - This is only really useful for direct connections.
+        /// - Relay connections and clients connected using a distributed authority network topology will not provide the client's actual endpoint information.
+        /// - For LAN topologies this should work as long as it is a direct connection and not a relay connection.
+        /// </remarks>
+        /// <param name="clientId">NGO client identifier to get endpoint information about.</param>
+        /// <returns><see cref="NetworkEndpoint"/></returns>
+        public NetworkEndpoint GetEndpoint(ulong clientId)
+        {
+            if (m_Driver.IsCreated && NetworkManager != null && NetworkManager.IsListening)
+            {
+                var (transportId, connectionExists) = NetworkManager.ConnectionManager.ClientIdToTransportId(clientId);
+                var networkConnection = ParseClientId(transportId);
+                if (connectionExists && m_Driver.GetConnectionState(networkConnection) == NetworkConnection.State.Connected)
+                {
 #if UTP_TRANSPORT_2_0_ABOVE
-        /// <summary>
-        /// Provides the <see cref="NetworkEndpoint"/> for the NGO client identifier specified.
-        /// </summary>
-        /// <remarks>
-        /// - This is only really useful for direct connections.
-        /// - Relay connections and clients connected using a distributed authority network topology will not provide the client's actual endpoint information.
-        /// - For LAN topologies this should work as long as it is a direct connection and not a relay connection.
-        /// </remarks>
-        /// <param name="clientId">NGO client identifier to get endpoint information about.</param>
-        /// <returns><see cref="NetworkEndpoint"/></returns>
-        public NetworkEndpoint GetEndpoint(ulong clientId)
-        {
-            if (m_Driver.IsCreated && NetworkManager != null && NetworkManager.IsListening)
-            {
-                var transportId = NetworkManager.ConnectionManager.ClientIdToTransportId(clientId);
-                var networkConnection = ParseClientId(transportId);
-                if (m_Driver.GetConnectionState(networkConnection) == NetworkConnection.State.Connected)
-                {
                     return m_Driver.GetRemoteEndpoint(networkConnection);
-                }
-            }
-            return new NetworkEndpoint();
-        }
 #else
-        /// <summary>
-        /// Provides the <see cref="NetworkEndpoint"/> for the NGO client identifier specified.
-        /// </summary>
-        /// <remarks>
-        /// - This is only really useful for direct connections.
-        /// - Relay connections and clients connected using a distributed authority network topology will not provide the client's actual endpoint information.
-        /// - For LAN topologies this should work as long as it is a direct connection and not a relay connection.
-        /// </remarks>
-        /// <param name="clientId">NGO client identifier to get endpoint information about.</param>
-        /// <returns><see cref="NetworkEndpoint"/></returns>
-        public NetworkEndpoint GetEndpoint(ulong clientId)
-        {
-            if (m_Driver.IsCreated && NetworkManager != null && NetworkManager.IsListening)
-            {
-                var transportId = NetworkManager.ConnectionManager.ClientIdToTransportId(clientId);
-                var networkConnection = ParseClientId(transportId);
-                if (m_Driver.GetConnectionState(networkConnection) == NetworkConnection.State.Connected)
-                {
                     return m_Driver.RemoteEndPoint(networkConnection);
+#endif
                 }
             }
             return new NetworkEndpoint();
         }
-#endif
 
 
         /// <summary>
@@ -1458,10 +1445,17 @@ namespace Unity.Netcode.Transports.UTP
                     // If the message is sent reliably, then we're over capacity and we can't
                     // provide any reliability guarantees anymore. Disconnect the client since at
                     // this point they're bound to become desynchronized.
+                    if (NetworkManager != null)
+                    {
+                        var (ngoClientId, isConnectedClient) = NetworkManager.ConnectionManager.TransportIdToClientId(clientId);
+                        if (isConnectedClient)
+                        {
+                            clientId = ngoClientId;
+                        }
 
-                    var ngoClientId = NetworkManager?.ConnectionManager.TransportIdToClientId(clientId) ?? clientId;
+                    }
                     Debug.LogError($"Couldn't add payload of size {payload.Count} to reliable send queue. " +
-                        $"Closing connection {ngoClientId} as reliability guarantees can't be maintained.");
+                        $"Closing connection {clientId} as reliability guarantees can't be maintained.");
 
                     if (clientId == m_ServerClientId)
                     {